refac: Move auth errors into own module
