+++ src/auth/mod.rs
@@ -1,6 +1,7 @@
 //! Authentication module.
 
 mod api_keys;
+mod errors;
 mod login;
 
 pub use login::login;
